feat: "Handled token expiration"

+++ server/api/blog/[id].delete.js
@@ -1,9 +1,10 @@
 import connectDB from '~/server/utils/db';
 import Blog from '~/server/models/Blog';
 import mongoose from 'mongoose';
+
 export default defineEventHandler(async(event)=>{
     await connectDB();
-
+     const { user } = event.context.auth
     const id = getRouterParam(event, "id")
     if(!mongoose.Types.ObjectId.isValid(id)){
         throw createError({
@@ -11,14 +12,24 @@ export default defineEventHandler(async(event)=>{
             statusMessage: 'Invalid blog id'
         })
     }
-    const blog = await Blog.findByIdAndDelete(id)
+  
+         let blog = await  Blog.findById(id)
+         if(!blog){
+            throw createError({
+                statusCode: 404,
+                statusMessage: 'Invalid blog id'
+            })
+        }
+         if(user._id.equals(blog.author)){
+             await Blog.findByIdAndDelete(id)
+         }else{
+            throw createError({
+                statusCode:401,
+                statusMessage:"Unauthorizedsssssssssssss"
+            })
+         }
 
-    if(!blog){
-        throw createError({
-            statusCode: 404,
-            statusMessage: 'Invalid blog id'
-        })
-    }
+    
     return{
         blog
     }
